requirements.txt
Flask-RESTful==0.3.9
marshmallow==4.0.0
flask-marshmallow==1.3.0
flask-cors==6.0.0
psycopg2-binary==2.9.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
pandas==2.3.1